+++ bactopia/__init__.py
@@ -1,6 +1,6 @@
 """Top-level package for Bactopia."""
 
-__version__ = '2.0.2'
+__version__ = '2.1.0'
 
 __all__ = [
     'const',