+++ bin/gh-actions/generate-nextflow-schema.py
@@ -4,7 +4,7 @@
 """
 
 PROGRAM = "generate-nextflow-schema.py"
-VERSION = "2.0.3"
+VERSION = "2.1.0"
 REPO = "bactopia"
 
 BACTOPIA_SCHEMA = "conf/schema/bactopia.json"
@@ -21,7 +21,6 @@ JSON_SCHEMAS = [
     "modules/local/bactopia/antimicrobial_resistance/params.json",
     "modules/local/bactopia/sequence_type/params.json",
     "modules/local/bactopia/blast/params.json",
-    "modules/local/bactopia/ariba_analysis/params.json",
     "conf/schema/generic.json"
 ]